+++ pkg/integrations/v2/eventhandler/eventhandler_test.go
@@ -5,6 +5,8 @@ import (
 	"testing"
 	"time"
 
+	v1 "k8s.io/api/core/v1"
+
 	"github.com/go-kit/log"
 	"github.com/stretchr/testify/require"
 )
@@ -21,4 +23,32 @@ func TestCacheLoad(t *testing.T) {
 	actualEvents, err := readInitEvent(cacheFile, l)
 	require.NoError(t, err, "Failed to parse last event from eventhandler cache file")
 	require.Equal(t, expectedEvents, actualEvents)
+}
+
+func TestExtractEventJson(t *testing.T) {
+	var eh = new(EventHandler)
+	eh.logFormat = logFormatJson
+	var event = new(v1.Event)
+	event.InvolvedObject = v1.ObjectReference{
+		Name: "test-object",
+	}
+	event.Message = "Event Message"
+
+	_, msg, err := eh.extractEvent(event)
+	require.NoError(t, err, "Failed to extract test event")
+	require.Equal(t, "{\"msg\":\"Event Message\",\"name\":\"test-object\"}", msg)
+}
+
+func TestExtractEventText(t *testing.T) {
+	var eh = new(EventHandler)
+	eh.logFormat = "text"
+	var event = new(v1.Event)
+	event.InvolvedObject = v1.ObjectReference{
+		Name: "test-object",
+	}
+	event.Message = "Event Message"
+
+	_, msg, err := eh.extractEvent(event)
+	require.NoError(t, err, "Failed to extract test event")
+	require.Equal(t, "name=test-object msg=\"Event Message\"", msg)
 }