+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Main (unreleased)
 
 - Add [godeltaprof](https://github.com/grafana/godeltaprof) profiling types (`godeltaprof_memory`, `godeltaprof_mutex`, `godeltaprof_block`) to `pyroscope.scrape` component
 - Integrations: make `udev` data path configurable in the `node_exporter` integration. (@sduranc)
+- Add `log_format` configuration to eventhandler integration. (@sadovnikov)
 
 - New Grafana Agent Flow components: